Exception handling and validations added.

+++ GameScope.Domain/CommandHandlers/DeleteGameCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using GameScope.Domain.Commands;
 using GameScope.Domain.Interfaces;
+using GameScope.Infra.Common.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -20,21 +21,35 @@ namespace GameScope.Domain.CommandHandlers
 
         public Task<bool> Handle(DeleteGameCommand request, CancellationToken cancellationToken)
         {
-            var game = _gameRepository.GetById(request.Id);
+            try
+            {
+                if (!request.IsValid())
+                {
+                    var error = request.ValidationResult.Errors[0];
 
-            if(game == null)
+                    throw new GameScopeException(error.ErrorCode, error.ErrorMessage);
+                }
+
+                var game = _gameRepository.GetById(request.Id);
+
+                if (game == null)
+                {
+                    throw new GameScopeException("game_not_found", $"You don't have a game with given information.");
+                }
+
+                if (game.UserId != request.UserId)
+                {
+                    throw new GameScopeException("unauthorized_game_delete", $"You don't have a permission to edit this game.");
+                }
+
+                _gameRepository.Remove(request.Id);
+
+                return Task.FromResult(_gameRepository.SaveChanges() > 0);
+            }
+            catch (Exception ex)
             {
                 return Task.FromResult(false);
             }
-
-            if(game.UserId != request.UserId)
-            {
-                return Task.FromResult(false);
-            }
-
-            _gameRepository.Remove(request.Id);
-
-            return Task.FromResult(_gameRepository.SaveChanges() > 0);
         }
     }
 }
